:truck: Move RandomStickersProvider to its own file
